move FavoriteSwitch into own component
